Consolidate item management into workspace-membership; use addItemToWorkspace in RoadmapPane
- Move moveItem from item-management.ts into workspace-membership.ts alongside addItemToWorkspace and removeItemFromWorkspace
- Delete item-management.ts
- Update RoadmapPane to use addItemToWorkspace store action instead of updateWorkspace
